解决部分编译问题

+++ Mcal/Platform/inc/Platform_Types.h
@@ -29,6 +29,7 @@
 /*==================================================================================================
 *                               SOURCE FILE VERSION INFORMATION
 ==================================================================================================*/
+// #include "IntCtrl_Lld_TypesDef.h"
 #ifdef MCAL_HAS_PLATFORM_TYPES_DEFINE
 #define PLATFORM_VENDOR_ID                      (180)
 #define PLATFORM_AR_RELEASE_MAJOR_VERSION       (4)
@@ -47,8 +48,8 @@ typedef signed char         sint8;         /**< -128 .. +127                   *
 typedef unsigned char       uint8;         /**< 0 .. 255                       */
 typedef signed short        sint16;        /**< -32768 .. +32767               */
 typedef unsigned short      uint16;        /**<  0 .. 65535                    */
-typedef signed long         sint32;        /**< -2147483648 .. +2147483647     */
-typedef unsigned long       uint32;        /**<  0 .. 4294967295               */
+typedef signed int         sint32;        /**< -2147483648 .. +2147483647     */
+typedef unsigned int       uint32;        /**<  0 .. 4294967295               */
 typedef signed long long    sint64;
 typedef unsigned long long  uint64;
